TINKERPOP-2438 Added AST preprocessing for scripts
There are sometimes things you want to know about what a script is doing before it goes to eval(). The immediate purpose is to extract timeouts passed using with() so that they can be applied to a request as an override to the server timeout. In this way, users who make the mistake of using this method (meant for remote bytecode requests initially) with script submission accidentally aren't left wondering why it didn't work as expected.